backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Trading Bot Backend - Backtest and Paper Trading System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.19",
    "fastapi>=0.121.2",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "ta-lib>=0.6.8",
    "uvicorn>=0.38.0",
    "vectorbt>=0.28.1",
    "websockets>=15.0.1",
    "yfinance>=0.2.40",
    "pyarrow>=18.1.0",
]
